// src/types/google.accounts.ts
export type ErrorCode =
  | 'invalid_request'
  | 'access_denied'
  | 'unauthorized_client'
  | 'unsupported_response_type'
  | 'invalid_scope'
  | 'server_error'
  | 'temporarily_unavailable';
// ...
export interface AccessToken {
  /** The access token of a successful token response. */
  access_token: string;

  /** The refresh token of a successful token response. */
  refresh_token?: string | null;

  /** The lifetime in seconds of the access token. */
  expires_in?: number;

  /** The hosted domain the signed-in user belongs to. */
  hd?: string;

  /** The prompt value that was used from the possible list of values specified by TokenClientConfig or OverridableTokenClientConfig */
  prompt?: string;

  /** The type of the token issued. */
  token_type?: string;

  /** A space-delimited list of scopes that are approved by the user. */
  scope?: string;

  /**  */
  id_token?: string;
  
  /** The string value that your application uses to maintain state between your authorization request and the response. */
  state?: string;

  /** A single ASCII error code. */
  error?: ErrorCode;

  /**	Human-readable ASCII text providing additional information, used to assist the client developer in understanding the error that occurred. */
  error_description?: string;

  /** A URI identifying a human-readable web page with information about the error, used to provide the client developer with additional information about the error. */
  error_uri?: string;

  /** The timestamp in seconds of the access token. */
  expired_timestamp?: number;

  /** as date */
  expired_date?: string;
}